added maps and moved data to memory

+++ src/main/java/net/crusadergames/bugwars/dto/request/GameRequest.java
@@ -16,5 +16,5 @@ public class GameRequest {
     private List<Long> scriptIds;
 
     @NotBlank
-    private Long mapId;
+    private Integer mapId;
 }
